moved calculate weights from radiation to abiotic tools
